requirements.txt
aiofiles==0.4.0
aniso8601==7.0.0
appdirs==1.4.3
async-exit-stack==1.0.1
async-generator==1.10
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
black==19.10b0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
decorator==4.4.1
dnspython==1.16.0
email-validator==1.0.5
fastapi==0.45.0
graphene==2.1.8
graphql-core==2.2.1
graphql-relay==2.0.1
gunicorn==19.9.0
h11==0.9.0
httptools==0.0.13
idna==2.8
importlib-metadata==1.3.0
ipdb==0.12.3
ipython==7.10.2
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==8.0.2
packaging==19.2
parso==0.5.2
passlib==1.7.2
pathspec==0.7.0
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.1
promise==2.3
prompt-toolkit==3.0.2
psycopg2==2.8.4
ptyprocess==0.6.0
py==1.8.0
pycparser==2.19
pydantic==1.2
Pygments==2.5.2
PyJWT==1.7.1
pyparsing==2.4.5
pytest==5.3.2
python-multipart==0.0.5
PyYAML==5.2
regex==2019.12.20
requests==2.22.0
Rx==1.6.1
six==1.13.0
slugify==0.0.1
SQLAlchemy==1.3.12
starlette==0.12.9
stripe==2.41.0
toml==0.10.0
traitlets==4.3.3
typed-ast==1.4.0
ujson==1.35
urllib3==1.25.7
uvicorn==0.10.9
uvloop==0.14.0
wcwidth==0.1.7
websockets==8.1
zipp==0.6.0
